Fixes
Added back the rescaling of weights to balance signal and background which was accidentally removed.

No longer removing pT from graphs before training.
The removal was there because the classifier training used to not work if the graphs had a pT attribute; now it works fine.

Brought back missing imports.