Fixes and improvements for histogram production
Fixes:
Using TH1F in make_histos.py meant that the bin count was limited to 2^24 = 16777216, which could be reached for some QCD histograms if using the full dataset.
A missing backslash in submit_slurm_make_histos.sh meant that the second of the two files in the JETM2_W_flatmass was not joined with the first one into a singla item and so was not being processed.

Speed improvements:
Filling histograms with arrays rather than with individual values in a loop.
QCD JZ slices processed in separate jobs.